requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
openpyxl==3.1.2
pandas==2.2.2
reportlab==4.1.0
pdfkit==1.0.0
APScheduler==3.10.4
loguru==0.7.2
python-dateutil==2.9.0
SQLAlchemy==2.0.30

python-telegram-bot==20.6
itsdangerous==2.2.0
pydantic-settings==2.2.1
fpdf==1.7.2
requests==2.31.0

